pyproject.toml
[project]
name = "vehicle-monitoring-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Leon Kushnir",email = "leon.kushnir15@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "vehicle_monitoring_system.settings"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
